QuantumPropagators

QuantumPropagators.initpropwrk — Method

Initialize a workspace for propagation.

wrk = initpropwrk(state, tlist, method=:auto, generator...; kwargs...)

The resulting wrk can be passed to propagate or propstep!.

Arguments

state: An exemplary state for the propagation (e.g., the initial state)

tlist: The time grid over which propagate will be called. Must include at least to points in order to determine the propagation time step to prepare. If the propagation will be over a tlist with a variable dt, the full tlist must be passed here.

generator: An exemplary (non-time-dependent) dynamical generator. For full generality (if method=:cheby), the given generator should have a spectral range sufficiently large to encompass the entire propagation. If given multiple times, a spectral envelope enclosing all the generators will be determined automatically. In this case, you should pass the generators with the extremal values of all the controls.

method: The propagation method to use. The default value of :auto attempts to choose the best method available, based on the properties of the given state, tlist, and generator. Alternative values are :cheby and :newton, and :expprop.

specrad_method: for method=:cheby, method to use for estimating the spectral radius, see QuantumPropagators.SpectralRange.specrange. Defaults to :auto.

tolerance: for method=:cheby, a tolerance factor for the estimated spectral radius. That is, Chebychev coefficients will be calculated for a spectral radius increased by the tolerance factor compared to the specral radius estimated for the generator.

All other kwargs are filtered and passed to the contructor for returned workspace, e.g. limit for method=:cheby or m_max for method=:newton. For method=:cheby, they are also passed to specrange.

QuantumPropagators.propagate — Method

Propagate a state over an entire time grid.

state_out = propagate(
    state, genfunc, tlist; method=:auto,
    backwards=false; storage=nothing, observables=(<store state>, ),
    hook=nothing, showprogress=false, control_parameters=nothing,
    in_place=true, kwargs...)

propagates state over the time grid in tlist, using piecewise-constant dynamical generators (Hamiltonians or Liouvillians) determined by genfunc, and returns the resulting propagated state. The propagation is performed by calling propstep! for every interval in tlist, or propstep if in_place=false.

For the i'th time interval, genfunc(tlist, i) must return the generator for that time interval. Generally, when approximating a time-continuous dynamical generator as piecewise-constant on the time grid, it should be evaluated at the midpoint of the interval. A possible exception is the first and last interval, which may be better evaluated at tlist[1] and tlist[end] to ensure exact boundary conditions like control fields that are exactly zero.

In addition to the two positional parameters indicating the time interval, genfunc will also receive the state (the input state for the propagation step), backwards, storage, observables, control_parameters, and init as keyword arguments.

The control_parameters are an optional array of floats with parameters for genfunc. This is required when propagate is used in the context of automatic differentiation (AD). E.g., the Zygote framework can automatically calculate gradients of a function control_parameters -> J_T, where J_T might be be a function of the overlap between a propagated state (returned by propagate) and a target state. Thus, the control_parameters must be explicit in propagate. Outside of an AD context, control_parameters are not generally required: they can be implicit in genfunc.

The remaining keyword arguments may be used for unusual equations of motion beyond the standard Schrödinger or Liouville-von-Neumann equation, e.g. state would enter the genfunc for a Gross–Pitaevskii equation. For standard equations of motion that do not use the additional parameters, it is best to capture the keyword arguments to genfunc with a definition like

genfunc(tlist, i; kwargs...) = ...

For valid propagation methods, see initpropwrk.

In general, there is no requirement that tlist has a constant time step, although some propagation methods (most notably cheby!) only support a uniform time grid.

If storage is given as an Array, it will be filled with data determined by the observables. The default "observable" results in the propagated states at every point in time being stored. The storage array should be created with init_storage. See its documentation for details.

The storage parameter may also be given as true, and a new storage array will be created internally with init_storage and returned instead of the propagated state:

data = propagate(
    state, genfunc, tlist; method=:auto
    backwards=false; storage=true, observables=observables,
    hook=nothing, showprogress=false, kwargs...)

If backwards is true, the input state is assumed to be at time tlist[end], and the propagation progresses backwards in time (with a negative time step dt). If storage is given, it will be filled back-to-front during the backwards propagation.

If hook is given as a callable, it will be called after each propagation step, as hook(state, generator, tlist, i, wrk, observables) where i is the index of the time interval on tlist covered by the propagation step (0 for the initial state, respectives lastindex(tlist) for the backward propagation). The hook is called before calculating any observables. Example usage includes writing data to file, or modifying state, e.g., removing amplitude from the lowest and highest level to mitigate "truncation error".

If showprogress is given as true, a progress bar will be shown for long-running propagationn. In order to customize the progress bar, showprogress may also be a function that receives length(tlist) and returns a ProgressMeter.Progress instance.

If in_place=false is given, the propagation avoids in-place operations by using propstep instead of propstep!. This is often required in the context of automatic differentiation (AD), e.g., with Zygote. That is, use in_place=false if propagate is called inside a function to be passed to Zygote.gradient, Zygote.pullback, or a similar function. In and AD context, storage and showprogress should not be used.

The propagate routine returns the propagated state at tlist[end], respectively tlist[1] if backwards=true, or a storage array with the stored states / observable data if storage=true.

QuantumPropagators.propstep! — Method

Perform a single propagation step in-place.

propstep!(state, generator, dt, wrk; kwargs...)

The propagation method is determined by wrk, see initpropwrk.

Generally, an in-place propagation will not be suitable for in the context of automatic differentiation. See propagate for a method that does not act in-place.

The kwargs are forwarded to the underlying method.

QuantumPropagators.propstep — Method

Perform a single propagation step and return the propagated state.

state_out = propstep(state, generator, dt, wrk; kwargs...)

The propagation method is determined by wrk, see initpropwrk.

Unlike propstep!, this method does not act in place, which generally makes it more suitable for automatic differentiation. However, there may be a performance penalty associated with the additional memory allocations.

The kwargs are forwarded to the underlying method.