QuantumControlBase

Documentation for QuantumControlBase.

QuantumControlBase.ControlProblem

QuantumControlBase.Objective

Base.adjoint

QuantumControlBase.blackman

QuantumControlBase.box

QuantumControlBase.discretize

QuantumControlBase.discretize_on_midpoints

QuantumControlBase.flattop

QuantumControlBase.getcontrols

QuantumControlBase.setcontrolvals

QuantumControlBase.setcontrolvals!

QuantumPropagators.initpropwrk

QuantumPropagators.propagate

QuantumControlBase.ControlProblem — Type

A full control problem with multiple objectives.

ControlProblem(
   objectives=<list of objectives>,
   pulse_options=<dict of controls to pulse options>,
   tlist=<time grid>,
   kwargs...
)

Note that the control problem can only be instantiated via keyword arguments.

The objectives are a list of Objective instances, each defining an initial state, a dynamical generator for the evolution of the state, and (optionally) a target for the evolution.

The pulse_options are a dictionary (IdDict) mapping controls that occur in the objectives to properties specific to the control method.

The tlist is the time grid on which the time evolution of the initial states of each objective should be propagated.

The remaining kwargs are keyword arguments that are passed directl to the optimal control method. These typically include e.g. the optimization functional.

The control problem is solved by finding a set of controls that simultaneously fulfill all objectives.

QuantumControlBase.Objective — Type

A single optimization objective.

Objective(
    initial_state=<intial state>,
    generator=<dynamical generator>,
    [target=<optional target state or specification>]
)

Note that the objective can only be instantiated via keyword arguments.

Base.adjoint — Method

Adjoint of an objective.

QuantumControlBase.blackman — Method

Blackman window shape.

blackman(t, t₀, T; a=0.16)

calculates

\[B(t; t_0, T) = \frac{1}{2}\left( 1 - a - \cos\left(2π \frac{t - t_0}{T - t_0}\right) + a \cos\left(4π \frac{t - t_0}{T - t_0}\right) \right)\,,\]

for a scalar t, with $a$ = 0.16.

See http://en.wikipedia.org/wiki/Window_function#Blackman_windows

A Blackman shape looks nearly identical to a Gaussian with a 6-sigma interval between t₀ and T. Unlike the Gaussian, however, it will go exactly to zero at the edges. Thus, Blackman pulses are often preferable to Gaussians.

QuantumControlBase.box — Method

Box shape (Theta-function).

box(t, t₀, T)

evaluates the Heaviside (Theta-) function $\Theta(t) = 1$ for $t_0 \le t \le T$; and $\Theta(t) = 0$ otherwise.

QuantumControlBase.discretize — Method

Evaluate control at every point of tlist.

values = discretize(control, tlist; via_midpoints=true)

discretizes the given control to a Vector of values defined on the points of tlist.

If control is a function, it will will first be evaluated at the midpoint of tlist, see discretize_on_midpoints, and then the values on the midpoints are converted to values on tlist. This discretization is more stable than directly evaluationg the control function at the values of tlist, and ensures that repeated round-trips between discretize and discretize_on_midpoints can be done safely, see the note in the documentation of discretize_on_midpoints.

The latter can still be achieved by passing via_midpoints=false. While such a direct discretization is suitable e.g. for plotting, but it is unsuitable for round-trips between discretize and discretize_on_midpoints (constant controls on tlist may result in a zig-zag on the intervals of tlist).

If control is a vector, it will be returned un-modified if it is of the same length as tlist. Otherwise, control must have one less value than tlist, and is assumed to be defined on the midpoins of tlist. In that case, discretize acts as the inverse of discretize_on_midpoints. See discretize_on_midpoints for how control values on tlist and control values on the intervals of tlist are related.

QuantumControlBase.discretize_on_midpoints — Method

Evaluate control at the midpoints of tlist.

values = discretize_on_midpoints(control, tlist)

discretizes the given control to a Vector of values on the midpoints of tlist. Hence, the resulting values will contain one less value than tlist.

If control is a vector of values defined on tlist (i.e., of the same length as tlist), it will be converted to a vector of values on the intervals of tlist. The value for the first and last "midpoint" will remain the original values at the beginning and end of tlist, in order to ensure exact bounary conditions. For all other midpoints, the value for that midpoint will be calculated by "un-averaging".

For example, for a control and tlist of length 5, consider the following diagram:

tlist index:       1   2   3   4   5
tlist:             ⋅   ⋅   ⋅   ⋅   ⋅   input values cᵢ (i ∈ 1..5)
                   |̂/ ̄ ̄ ̂\ / ̂\ / ̂ ̄ ̄\|̂
midpoints:         x     x   x     x   output values pᵢ (i ∈ 1..4)
midpoints index:   1     2   3     4

We will have $p₁=c₁$ for the first value, $p₄=c₅$ for the last value. For all other points, the control values $cᵢ = \frac{p_{i-1} + p_{i}}{2}$ are the average of the values on the midpoints. This implies the "un-averaging" for the midpoint values $pᵢ = 2 c_{i} - p_{i-1}$.

Note

An arbitrary input control array may not be compatible with the above averaging formula. In this case, the conversion will be "lossy" (discretize will not recover the original control array; the difference should be considered a "discretization error"). However, any further round-trip conversions between points and intervals are bijective and preserve the boundary conditions. In this case, the discretize_on_midpoints and discretize methods are each other's inverse. This also implies that for an optimal control procedure, it is safe to modify midpoint values. Modifying the the values on the time grid directly on the other hand may accumulate discretization errors.

If control is a vector of one less length than tlist, it will be returned unchanged, under the assumption that the input is already properly discretized.

If control is a function, the function will be directly evaluated at the midpoints marked as x in the above diagram..

QuantumControlBase.flattop — Method

Flat shape (one) with a switch-on/switch-off from zero.

flattop(t; t₀, T, t_rise, t_fall=t_rise, func=:blackman)

evaluates a shape function that starts at 0 at $t=t₀$, and ramps to to 1 during the t_rise interval. The function then remains at value 1, before ramping down to 0 again during the interval t_fall before T. For $t < t₀$ and $t > T$, the shape is zero.

The default switch-on/-off shape is half of a Blackman window (see blackman).

For func=:sinsq, the switch-on/-off shape is a sine-squared curve.

QuantumControlBase.getcontrols — Method

Extract a Tuple of controls.

controls = getcontrols(generator)

extracts the controls from a single dynamical generator.

controls = getcontrols(objectives)

extracts the controls from a list of objectives (i.e., from each objective's generator)

In either case, controls that occur multiple times, either in a single generator, or throughout the different objectives, will occur only once in the result.

By default, assumes that any generator is a nested Tuple, e.g. (H0, (H1, ϵ1), (H2, ϵ2), ...) and extracts (ϵ1, ϵ2)

Each control must be a valid argument for discretize.

QuantumControlBase.setcontrolvals! — Method

In-place version of setcontrolvals.

QuantumControlBase.setcontrolvals — Method

Construct G by plugging values into a general generator.

G = setcontrolvals(generator, vals_dict)
setcontrolvals!(G, generator, vals_dict)

evaluates the specific dynamical generator G by plugging in values into the general generator according to vals_dict.

The vals_dict is a dictionary (IdDict) mapping controls as returned by getcontrols(generator) to values.

QuantumPropagators.initpropwrk — Method

wrk = initpropwrk(obj, tlist; method=:auto, kwargs...)

initializes a workspace for the propagation of a control Objective.

Note that method must be given as a keyword argument.

QuantumPropagators.propagate — Function

Propagate the initial state of a control objective.

propagate(obj, tlist, method=:auto; controls_map=IdDict(), kwargs...)

propagates obj.initial_state under the dynamics described by obj.generator.

The optional dict control_map may be given to replace the controls in obj.generator (as obtained by getcontrols) with custom functions or vectors, e.g. with the controls resulting from optimization.

All kwargs are forwarded to QuantumPropagators.propagate.